Way too many changes for one commit: Organisation + blockquote stripes
Now all the clientside scripts are bundled with compilist and uploaded to
Tumblr's theme assets utility, rather than pasted inline. style.styl is still
embedded in the theme for convenience's sake.

This changeset also makes following long reblog conversations easier: The
blockquotes will be made narrower once a conversation is long and nested
enough, and be striped in different colours corresponding to the users posting
each message in the chain. This is achieved using a new script in
src/features/blockquotes.