Set min-width on the main element so it doesn't "snap" between sizes
Previously, as the window width shrank, the main element would reduce in size
constantly (corresponding to its 70% width), until the window hit 48 ems wide,
at which point the first responsive breakpoint kicks in and the main column
jumps back to 100% width.

Now the main element has a min-width that matches the first breakpoint at 48
ems, so it stops shrinking at the "right" spot for the breakpoint to kick in
and hence doesn't suddenly grow bigger again at a certain point.